pyproject.toml
[project]
name = "agno-medical-assistant"
version = "0.1.0"
description = "AI Medical Assistant System with Triage, RAG, and OCR capabilities"
authors = [
    {name = "Davi Castro Samora", email = "davicastrosamora@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # Core AI Framework
    "agno>=0.1.0",

    # LLM Providers
    "groq>=0.4.0",
    "openai>=1.10.0",

    # Observability
    "langfuse>=2.0.0",

    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",

    # Vector Database & RAG
    "duckdb>=0.10.0",
    "sentence-transformers>=2.3.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",

    # OCR & Document Processing
    "pymupdf>=1.23.0",
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
    "pdf2image>=1.16.0",

    # Data Processing
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",

    # Utilities
    "httpx>=0.26.0",
    "aiofiles>=23.2.0",
    "tenacity>=8.2.0",
    "loguru>=0.7.0",

    # Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
